refactor: split lexer literals
